requirements.txt
pydantic==1.8.1
aiofiles==0.6.0
autoscriber==1.0.3
fastapi==0.63.0
mysql-connector-python==8.0.23
uvicorn==0.13.4
